Always over-commit ffi_closure.tramp to 128 bytes.
This is a dirty, evil, laundry-eating hack. The thing is, FFI_TRAMPOLINE_SIZE
is a macro in libffi and the developers tend to change it without any regard
for languages that can't rely on the macros. Thus, to maintain some sort of
ABI compatibility, we'll simply over-commit. This has the fun side-effect of
preventing us from sanely using any fields on ffi_closure, but luckily we
don't need to do this at the moment.

See: https://github.com/atgreen/libffi/issues/10